+++ src/shared/globals.ts
@@ -4,6 +4,8 @@ import { GetSiteOptions, getStoredChart } from '#shared/storage'
 import { GenerateDefaultChart } from '#shared/chart'
 import { GenerateDefaultSiteOptions } from '#shared/misc'
 
+// TODO: look into adding Pinia for state management
+
 // pull from storage or generate site options and chartstate
 const storedOptions = GetSiteOptions()
 const loadOptions = storedOptions ? storedOptions : GenerateDefaultSiteOptions()
@@ -11,7 +13,7 @@ const loadOptions = storedOptions ? storedOptions : GenerateDefaultSiteOptions()
 const storedChart = getStoredChart(loadOptions.currentChart)
 const loadChartState = storedChart ? storedChart : GenerateDefaultChart()
 
-// Changed back to ref() from reactive() for more proper value changing
 export const GlobalSiteOptions: Ref<SiteOptions> = ref(loadOptions)
 export const GlobalChartState: Ref<ChartState> = ref(loadChartState)
 export const StoredChartNames = ref([''])
+export const selectedChartTitle = ref('')